Checked tests are all still passing

# Conflicts:
#	src/main/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountPaymentTermsService.java
#	src/test/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountServicePaymentCardTest.java
#	src/test/java/uk/gov/hmcts/opal/service/opal/OpalDefendantAccountServiceTest03.java

+++ src/main/java/uk/gov/hmcts/opal/dto/legacy/GetMajorCreditorAccountHistoryLegacyResponse.java
@@ -8,6 +8,7 @@ import jakarta.xml.bind.annotation.XmlElementWrapper;
 import jakarta.xml.bind.annotation.XmlRootElement;
 import jakarta.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -26,7 +27,7 @@ import uk.gov.hmcts.opal.util.LocalDateTimeAdapter;
 public class GetMajorCreditorAccountHistoryLegacyResponse implements ToXmlString {
 
     @XmlElement(name = "version")
-    private Long version;
+    private BigInteger version;
 
     @JsonProperty("history_items")
     @XmlElementWrapper(name = "history_items")
